requirements.txt
# CompanySearch backend + ingestion
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opensearch-py>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
spacy>=3.7.0
kaggle>=1.6.0
